Fix a layout bug
When using the app on an iPhone, the bottom part of the featured match
is often obscured.  Which is a shame because that portrait totally looked
just like the Mrs.

This glitch manifested because of a maximum-height for a layout block
that collaposes from horizontal to vertical stacking at small window
widths. We can safely stash that max-height declaration in a media-query
with a min-width and we should be golden.

DISCLAIMER: I haven't actually tested this but it looks plausible